+++ rotkehlchen/tests/api/test_addressbook.py
@@ -604,11 +604,7 @@ def test_names_compilation(rotkehlchen_api_server: 'APIServer') -> None:
                     name='Cody',
                     address=address_cody,
                     blockchain=SupportedBlockchain.ETHEREUM,
-                ), AddressbookEntry(
-                    name='rotki global db',
-                    address=address_rotki,
-                    blockchain=SupportedBlockchain.ETHEREUM,
-                ),
+                ),  # address_rotki is already added
             ],
         )
     global_addressbook_addresses = [
@@ -616,7 +612,7 @@ def test_names_compilation(rotkehlchen_api_server: 'APIServer') -> None:
         OptionalChainAddress(address_rotki, SupportedBlockchain.ETHEREUM),
     ]
     global_addressbook_expected = {
-        AddressbookEntry(address=address_rotki, blockchain=SupportedBlockchain.ETHEREUM, name='rotki global db'),  # noqa: E501
+        AddressbookEntry(address=address_rotki, blockchain=SupportedBlockchain.ETHEREUM, name='rotki.eth'),  # noqa: E501
         AddressbookEntry(address=address_cody, blockchain=SupportedBlockchain.ETHEREUM, name='Cody'),  # noqa: E501
     }
     response = names_request(global_addressbook_addresses)
@@ -656,11 +652,7 @@ def test_names_compilation(rotkehlchen_api_server: 'APIServer') -> None:
                     name='Rose',
                     address=address_rose,
                     blockchain=SupportedBlockchain.ETHEREUM,
-                ), AddressbookEntry(
-                    name='rotki private db',
-                    address=address_rotki,
-                    blockchain=SupportedBlockchain.ETHEREUM,
-                ),
+                ),  # address_rotki is already added
             ],
         )
     private_addressbook_addresses = [
@@ -670,7 +662,7 @@ def test_names_compilation(rotkehlchen_api_server: 'APIServer') -> None:
         OptionalChainAddress(address_rotki, SupportedBlockchain.ETHEREUM),  # rotki.eth
     ]
     private_addressbook_expected = {
-        AddressbookEntry(address=address_rotki, blockchain=SupportedBlockchain.ETHEREUM, name='rotki private db'),  # noqa: E501
+        AddressbookEntry(address=address_rotki, blockchain=SupportedBlockchain.ETHEREUM, name='rotki label'),  # noqa: E501
         AddressbookEntry(address=address_tylor, blockchain=SupportedBlockchain.ETHEREUM, name='Tylor'),  # noqa: E501
         AddressbookEntry(address=address_cody, blockchain=SupportedBlockchain.ETHEREUM, name='Cody'),  # noqa: E501
         AddressbookEntry(address=address_rose, blockchain=SupportedBlockchain.ETHEREUM, name='Rose'),  # noqa: E501